+++ src/org/camunda/latera/bss/connectors/MailSender.groovy
@@ -12,6 +12,7 @@ import javax.mail.internet.MimeBodyPart
 import javax.mail.internet.MimeMessage
 import javax.mail.internet.MimeMultipart
 import javax.mail.internet.MimeUtility
+import static org.camunda.latera.bss.utils.StringUtil.isEmpty
 import static org.camunda.latera.bss.utils.ListUtil.firstNotNull
 
 class MailSender {
@@ -23,14 +24,16 @@ class MailSender {
   private Message    message
   private Multipart  multipart
   private Properties props
+  private Transport  transport
 
   MailSender(DelegateExecution execution) {
     def ENV       = System.getenv()
 
-    this.host     =  execution.getVariable('smtpHost')     ?: ENV['SMTP_HOST']
-    this.port     = (execution.getVariable('smtpPort')     ?: ENV['SMTP_PORT'] ?: 587).toInteger()
-    this.user     =  execution.getVariable('smtpUser')     ?: ENV['SMTP_USER']
-    this.password =  execution.getVariable('smtpPassword') ?: ENV['SMTP_PASSWORD']
+    this.host      =  execution.getVariable('smtpHost')     ?: ENV['SMTP_HOST']
+    this.port      = (execution.getVariable('smtpPort')     ?: ENV['SMTP_PORT'] ?: 587).toInteger()
+    this.user      =  execution.getVariable('smtpUser')     ?: ENV['SMTP_USER']
+    this.password  =  execution.getVariable('smtpPassword') ?: ENV['SMTP_PASSWORD']
+    this.transport =  session.getTransport('smtp')
 
     Boolean auth = Boolean.valueOf(firstNotNull([
       execution.getVariable('smtpAuth'),
@@ -67,13 +70,30 @@ class MailSender {
       this.props.put('mail.smtp.ssl.trust', host)
     }
 
-    this.session   = Session.getDefaultInstance(props, null)
-    this.message   = new MimeMessage(session)
-    this.multipart = new MimeMultipart()
-    this.setFrom(user)
+    this.session = Session.getDefaultInstance(props, null)
+    this.newMessage()
   }
 
-  MailSender getHost() {
+  MailSender connect() {
+    if (!this.transport.isConnected()) {
+      if (isEmpty(this.host)) {
+        throw new Exception("Empty host name!")
+      }
+      if (isEmpty(this.port)) {
+        throw new Exception("Empty port number!")
+      }
+      if (isEmpty(this.user)) {
+        throw new Exception("Empty user name!")
+      }
+      if (isEmpty(this.password)) {
+        throw new Exception("Empty password!")
+      }
+      this.transport.connect(this.host, this.port, this.user, this.password)
+    }
+    return this
+  }
+
+  String getHost() {
     return this.host
   }
 
@@ -83,7 +103,7 @@ class MailSender {
     return this
   }
 
-  MailSender getPort() {
+  Integer getPort() {
     return this.port
   }
 
@@ -97,7 +117,7 @@ class MailSender {
     return setPort(port.toInteger())
   }
 
-  MailSender getUser() {
+  String getUser() {
     return this.user
   }
 
@@ -114,32 +134,39 @@ class MailSender {
   }
 
   MailSender setFrom(CharSequence from) {
-    message.setFrom(new InternetAddress(from.toString()))
+    this.message.setFrom(new InternetAddress(from.toString()))
+    return this
+  }
+
+  MailSender newMessage() {
+    this.message   = new MimeMessage(session)
+    this.multipart = new MimeMultipart()
+    setFrom(user)
     return this
   }
 
   MailSender addRecipient(CharSequence recipient, Message.RecipientType type = Message.RecipientType.TO) {
-    message.addRecipient(type, new InternetAddress(recipient.toString()))
+    this.message.addRecipient(type, new InternetAddress(recipient.toString()))
     return this
   }
 
   MailSender setSubject(CharSequence subject) {
-    message.setSubject(subject.toString())
+    this.message.setSubject(subject.toString())
     return this
   }
 
   MailSender addTextPart(CharSequence body) {
     MimeBodyPart part = new MimeBodyPart()
     part.setText(body.toString())
-    multipart.addBodyPart(part)
+    this.multipart.addBodyPart(part)
     return this
   }
 
-  MailSender addFile(CharSequence filename, Object datasource){
+  MailSender addFile(CharSequence name, Object datasource) {
     MimeBodyPart part = new MimeBodyPart()
     part.setDataHandler(new DataHandler(datasource, 'application/octet-stream'))
-    part.setFileName(filename.toString())
-    multipart.addBodyPart(part)
+    part.setFileName(MimeUtility.encodeText(name, 'UTF-8', null))
+    this.multipart.addBodyPart(part)
     return this
   }
 
@@ -148,19 +175,23 @@ class MailSender {
     URL url = new URL(urlStr)
     part.setDataHandler(new DataHandler(url))
     part.setFileName(MimeUtility.encodeText(name, 'UTF-8', null))
-    multipart.addBodyPart(part)
+    this.multipart.addBodyPart(part)
     return this
   }
 
-  void send(){
-    Transport transport = session.getTransport('smtp')
-    transport.connect(host, port, user, password)
+  Boolean send() {
+    Boolean result = true
+    connect()
     try {
-      message.setContent(multipart)
-      transport.sendMessage(message, message.getAllRecipients())
-    }
-    finally {
+      this.message.setContent(this.multipart)
+      this.transport.sendMessage(this.message, this.message.getAllRecipients())
+    } catch (Exception e) {
+      logger.error(e)
+      result = false
+    } finally {
       transport.close()
     }
+
+    return result
   }
 }