+++ src/org/camunda/latera/bss/connectors/hid/hydra/Ref.groovy
@@ -133,7 +133,7 @@ trait Ref {
   }
 
   Number getRefIdByCode(CharSequence code) {
-    def id = Constants.getContstantByCode(code)
+    def id = Constants.getConstantByCode(code)
     if (id) {
       return id
     }
@@ -148,14 +148,15 @@ trait Ref {
     ]
     id = hid.getTableFirst(getRefsTable(), 'n_ref_id', where)
     if (id == null) {
-      id = toIntSafe(hid.queryFirst("SELECT SYS_CONTEXT('CONST', '${code}') FROM DUAL")?.getAt(0))
-      if (id != null) {
-        code = getRefCodeById(id)
-      }
-      return id
-    } else {
+      try {
+        id = toIntSafe(hid.queryFirst("SELECT SYS_CONTEXT('CONST', '${code}') FROM DUAL")?.getAt(0))
+      } catch (Exception e) {}
+    }
+
+    if (id != null) {
       return RefCache.instance.putAndGet(code, id)
     }
+    return null
   }
 
   Number getRefIdByName(CharSequence name) {
@@ -166,7 +167,7 @@ trait Ref {
   }
 
   String getRefCode(def id) {
-    String code = Constants.getContstantCode(id)
+    String code = Constants.getConstantCode(id)
     if (code) {
       return code
     }